Because of how the workers work in android It wont be able to make requests more often than 15 minutes. Added notification in case of error in the requests. Change layout and added some more labels Added worker I guess that's it for now

+++ app/src/main/java/dev/abrantes/monitor/infrastructure/Response.kt
@@ -12,5 +12,5 @@ data class Response(
     @NotNull val uri: String,
     @NotNull val dat: String,
     @NotNull val status: Int,
-    @ColumnInfo(name = "request_time") @NotNull val requestTime: Int
+    @ColumnInfo(name = "request_time") @NotNull val requestTime: Long
 )
